Migrate Team members list from oacore/next
Converts `members` property from the `oacore/next/data/team.yml`
to a list of markdown files and stores them separately in this
content repository.

Migrates images along with the textual data.

Adds order to the team members, what fixes sorting of the team members
and allows to put the founder first.